feat: move some methods to MethodContext  from AstBuilder
