+++ src/main/java/kalang/compiler/compile/AstBuilder.java
@@ -28,6 +28,8 @@ import javax.annotation.Nullable;
 import java.lang.reflect.Modifier;
 import java.util.*;
 
+import static kalang.compiler.compile.MethodContext.*;
+
 /**
  *  build ast from antlr parse tree
  * 
@@ -123,11 +125,6 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
     //private VarTable<VarObject,NullableKind> assignedNullables = new VarTable();
     
     //private VarTable<VarObject,Boolean> mustNullorNonnull = new VarTable();
-    
-    private static final int NULLSTATE_MUST_NULL = 0,
-            NULLSTATE_MUST_NONNULL = 1,
-            NULLSTATE_UNKNOWN = 2,
-            NULLSTATE_NULLABLE = 3;
 
     //private final HashMap<MethodNode,BlockStmtContext> methodBodys = new HashMap<>();
 
@@ -148,100 +145,6 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
     private KalangParser parser;
     
     private final CompilationUnit compilationUnit;
-    
-    private void newOverrideTypeStack(){
-        methodCtx.overrideTypes = new VarTable(methodCtx.overrideTypes);
-    }
-    
-    private void popOverrideTypeStack(){
-        methodCtx.overrideTypes = methodCtx.overrideTypes.getParent();
-    }
-    
-    private void removeOverrideType(ExprNode expr){
-        VarObject key = getOverrideTypeKey(expr);
-        if(key!=null) methodCtx.overrideTypes.remove(key, true);
-    }
-    
-    @Nullable
-    private VarObject getOverrideTypeKey(ExprNode expr){
-        VarObject key ;
-        //It isn't supported to override type of field because it is not safe
-        if(expr instanceof VarExpr){
-            key = ((VarExpr) expr).getVar();
-        }else if(expr instanceof ParameterExpr){
-            key = ((ParameterExpr) expr).getParameter();
-        }else{
-            key = null;
-        }
-        return key;
-    }
-    
-    private void changeTypeTemporarilyIfCould(ExprNode expr,Type type){
-        VarObject key = getOverrideTypeKey(expr);
-        if(key!=null){
-            methodCtx.overrideTypes.put(key, type);
-        }
-    }
-    
-    private void onNull(ExprNode expr,boolean onTrue,boolean isEQ){
-        boolean mustNull = (onTrue && isEQ) || (!onTrue && !isEQ);
-        VarObject key = this.getOverrideTypeKey(expr);
-        if(key!=null){
-            methodCtx.nullState.put(key,mustNull ? NULLSTATE_MUST_NULL : NULLSTATE_MUST_NONNULL);
-        }
-    }
-    
-    protected void onIf(ExprNode expr,boolean onTrue){
-        if(expr instanceof InstanceOfExpr && onTrue){
-            InstanceOfExpr ie = (InstanceOfExpr) expr;
-            changeTypeTemporarilyIfCould(ie.getExpr(),Types.getClassType(ie.getTarget().getReferencedClassNode()));
-        } else if(expr instanceof CompareExpr) {
-            CompareExpr ce = (CompareExpr) expr;
-            ExprNode e1 = ce.getExpr1();
-            ExprNode e2 = ce.getExpr2();
-            boolean isEQ = ce.getOperation().equals(CompareExpr.OP_EQ);
-            if (e1.getType().equals(Types.NULL_TYPE)) {
-                onNull(e2, onTrue, isEQ);
-            } else if (e2.getType().equals(Types.NULL_TYPE)) {
-                onNull(e1, onTrue, isEQ);
-            }
-        } else if (expr instanceof StaticInvokeExpr) {
-            StaticInvokeExpr sie = (StaticInvokeExpr) expr;
-            ExprNode[] args = sie.getArguments();
-            if (args==null || args.length != 2) {
-                return;
-            }
-            String invokeClass = sie.getInvokeClass().getReferencedClassNode().name;
-            if (!Objects.class.getName().equals(invokeClass)) {
-                return;
-            }
-            String methodName = sie.getMethod().getName();
-            if (!"equals".equals(methodName) && !"deepEquals".equals(methodName)) {
-                return;
-            }
-            if (Types.NULL_TYPE.equals(args[0].getType())) {
-                onNull(args[1],onTrue,true);
-            } else if (Types.NULL_TYPE.equals(args[1].getType())) {
-                onNull(args[0],onTrue,true);
-            }
-        } else if(expr instanceof UnaryExpr){
-            onIf(((UnaryExpr) expr).getExpr(),!onTrue);
-        } else if(expr instanceof LogicExpr){
-            LogicExpr le = (LogicExpr) expr;
-            if(le.getOperation().equals(LogicExpr.OP_LOGIC_AND)){
-                if(onTrue){
-                    onIf(le.getExpr1(),true);
-                    onIf(le.getExpr2(),true);
-                }
-            }else if(le.getOperation().equals(LogicExpr.OP_LOGIC_OR)){
-                if(!onTrue){
-                    onIf(le.getExpr1(),false);
-                    onIf(le.getExpr2(),false);
-                }
-            }
-        }
-        
-    }
     
     public ParserRuleContext getParseTree(){
         return compilationContext;
@@ -378,22 +281,14 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
         return bs;
     }
     
-    private void newFrame(){
-        methodCtx.newFrame();
-    }
-    
-    private void popFrame(){
-        methodCtx.popFrame();
-    }
-    
     BlockStmt newBlock(){
         BlockStmt bs = new BlockStmt();
-        this.newFrame();
+        methodCtx.newFrame();
         return bs;
     }
     
     void popBlock(){
-        this.popFrame();
+        methodCtx.popFrame();
     }
 
     @Nonnull
@@ -490,14 +385,14 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
     public AstNode visitQuestionExpr(KalangParser.QuestionExprContext ctx) {
         List<Statement> stmts = new LinkedList<>();
         ExprNode conditionExpr = (ExprNode) visit(ctx.expression(0));
-        newOverrideTypeStack();
-        onIf(conditionExpr,true);
+        methodCtx.newOverrideTypeStack();
+        methodCtx.onIf(conditionExpr,true);
         ExprNode trueExpr = (ExprNode) visit(ctx.expression(1));
-        popOverrideTypeStack();
-        newOverrideTypeStack();
-        onIf(conditionExpr,false);
+        methodCtx.popOverrideTypeStack();
+        methodCtx.newOverrideTypeStack();
+        methodCtx.onIf(conditionExpr,false);
         ExprNode falseExpr = (ExprNode)  visit(ctx.expression(2));
-        popOverrideTypeStack();
+        methodCtx.popOverrideTypeStack();
         Type trueType = trueExpr.getType();
         Type falseType  = falseExpr.getType();
         Type type;
@@ -626,62 +521,31 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
         BlockStmt falseBody = null;
         VarTable<VarObject,Integer> trueAssigned,falseAssigned;
         this.methodCtx.nullState = trueAssigned = this.methodCtx.nullState.newStack();
-        newOverrideTypeStack();
-        onIf(expr, true);
+        methodCtx.newOverrideTypeStack();
+        methodCtx.onIf(expr, true);
         if (ctx.trueStmt != null) {
             trueBody=requireBlock(ctx.trueStmt);
         }
-        popOverrideTypeStack();
+        methodCtx.popOverrideTypeStack();
         this.methodCtx.nullState = this.methodCtx.nullState.popStack();
         boolean trueReturned = this.methodCtx.returned;
         this.methodCtx.returned = false;
         this.methodCtx.nullState = falseAssigned = this.methodCtx.nullState.newStack();
-        newOverrideTypeStack();
-        onIf(expr,false);
+        methodCtx.newOverrideTypeStack();
+        methodCtx.onIf(expr,false);
         if (ctx.falseStmt != null) {
             falseBody=requireBlock(ctx.falseStmt);
         }
-        popOverrideTypeStack();
+        methodCtx.popOverrideTypeStack();
         this.methodCtx.nullState = this.methodCtx.nullState.popStack();
-        handleMultiBranchedAssign(trueAssigned.vars(),falseAssigned.vars());
+        methodCtx.handleMultiBranchedAssign(trueAssigned.vars(),falseAssigned.vars());
         boolean falseReturned = this.methodCtx.returned;
-        if(trueReturned) onIf(expr,false);
-        if(falseReturned) onIf(expr,true);
+        if(trueReturned) methodCtx.onIf(expr,false);
+        if(falseReturned) methodCtx.onIf(expr,true);
         this.methodCtx.returned = falseReturned && trueReturned;
         IfStmt ifStmt = new IfStmt(expr,trueBody,falseBody);
         mapAst(ifStmt,ctx);
         return ifStmt;
-    }
-    
-    private void handleMultiBranchedAssign(Map<VarObject,Integer>... assignedTable){
-        if(assignedTable.length<2){
-            throw Exceptions.illegalArgument(assignedTable);
-        }
-        HashMap<VarObject,Integer> ret = new HashMap();
-        ret.putAll(assignedTable[0]);
-        for(int i=1;i<assignedTable.length;i++){
-            Map<VarObject,Integer> other = assignedTable[i];
-            for(Map.Entry<VarObject,Integer> e:ret.entrySet()){
-                Integer oneNullable = e.getValue();
-                Integer otherNullable = other.get(e.getKey());
-                if(oneNullable.equals(otherNullable)) continue;
-                if(otherNullable==null) ret.remove(e.getKey());
-                else{
-                    int ns;
-                    if(
-                            (oneNullable.equals(NULLSTATE_MUST_NONNULL) && otherNullable.equals(NULLSTATE_UNKNOWN))                           || (otherNullable.equals(NULLSTATE_MUST_NONNULL) && oneNullable.equals(NULLSTATE_UNKNOWN))
-                            ){
-                        ns = NULLSTATE_UNKNOWN;
-                    }else{
-                        ns = NULLSTATE_NULLABLE;
-                    }
-                    ret.put(e.getKey(),ns);
-                }
-            }
-        }
-        for(Map.Entry<VarObject,Integer> e:ret.entrySet()){
-            this.methodCtx.nullState.put(e.getKey(), e.getValue());
-        }
     }
 
     protected ExprNode visitExpression(ExpressionContext expression) {
@@ -1017,7 +881,7 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
                 AssignExpr aexpr = new AssignExpr(toExpr,from);
                 mapAst(aexpr, ctx);
                 //TODO remove override information before assign
-                onAssign(toExpr,from);
+                methodCtx.onAssign(toExpr,from);
                 expr = aexpr;
             }else{
                 AstBuilder.this.handleSyntaxError("unsupported assign statement",ctx);
@@ -1025,45 +889,6 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
             }
         }
         return expr;
-    }
-    
-    private int getNullState(NullableKind nullable){
-        int ns;
-        if(nullable==NullableKind.NONNULL){
-            ns = NULLSTATE_MUST_NONNULL;
-        }else if(nullable==NullableKind.NULLABLE){
-            ns = NULLSTATE_NULLABLE;
-        }else if(nullable==NullableKind.UNKNOWN){
-            ns = NULLSTATE_UNKNOWN;
-        }else{
-            throw Exceptions.unexceptedValue(nullable);
-        }
-        return ns;
-    }
-    
-    private void onAssign(ExprNode to,ExprNode expr){
-        removeOverrideType(to);
-        if(to instanceof VarExpr){
-            ((VarExpr)to).removeOverrideType();
-        }else if(to instanceof ParameterExpr){
-            ((ParameterExpr) to).removeOverrideType();
-        }
-        VarObject key = getOverrideTypeKey(to);
-        if(key!=null){
-            Type toType = to.getType();
-            if(toType instanceof ObjectType){
-                Type type = expr.getType();
-                int ns;
-                if(Types.NULL_TYPE.equals(type)){
-                    ns = NULLSTATE_MUST_NULL;
-                }else if(type instanceof ObjectType){
-                    ns = getNullState(((ObjectType) type).getNullable());
-                }else{
-                    throw Exceptions.unexceptedValue(type);
-                }
-                methodCtx.nullState.put(key, ns);
-            }
-        }
     }
     
     @Nullable
@@ -1388,40 +1213,24 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
     }
      
     private boolean isDefindedId(String id){
-        return this.getNamedLocalVar(id)!=null 
-                || this.getNamedParameter(id) != null;
-    }
-    
-    @Nullable
-    private ParameterNode getNamedParameter(String name){
-        if (methodCtx != null) {
-            for (ParameterNode p : methodCtx.method.getParameters()) {
-                if (p.getName().equals(name)) {
-                    return p;
-                }
-            }
-        }
-        return null;
-    }
-    
-    @Nullable
-    private LocalVarNode getNamedLocalVar(String name){
-        return methodCtx!=null ? this.methodCtx.varTables.get(name) : null;
+        return methodCtx!=null &&
+                (methodCtx.getNamedLocalVar(id)!=null
+                || methodCtx.getNamedParameter(id) != null);
     }
 
     @Nullable
     private AstNode getNodeById(@Nonnull String name,@Nullable Token token) {
         //find local var
-        LocalVarNode var = this.getNamedLocalVar(name);
+        LocalVarNode var = methodCtx!=null ? methodCtx.getNamedLocalVar(name) : null;
         if(var!=null){
-            VarExpr ve = new VarExpr(var,getVarObjectType(var));
+            VarExpr ve = new VarExpr(var,methodCtx.getVarObjectType(var));
             if(token!=null) mapAst(ve, token);
             return ve;
         }
         //find parameters
-        ParameterNode paramNode = this.getNamedParameter(name);
+        ParameterNode paramNode = methodCtx==null ? null : methodCtx.getNamedParameter(name);
         if(paramNode!=null){
-            ParameterExpr ve = new ParameterExpr(paramNode,this.getVarObjectType(paramNode));
+            ParameterExpr ve = new ParameterExpr(paramNode,methodCtx.getVarObjectType(paramNode));
             if(token!=null) mapAst(ve, token);
             return ve;
         }
@@ -1566,7 +1375,7 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
         List<CatchBlock> tryCatchBlocks = new LinkedList<>();
         if (ctx.catchTypes != null) {
             for (int i = 0; i < ctx.catchTypes.size(); i++) {
-                this.newFrame();
+                methodCtx.newFrame();
                 this.methodCtx.returned = false;
                 String vName = ctx.catchVarNames.get(i).getText();
                 String vType = ctx.catchTypes.get(i).getText();
@@ -1576,7 +1385,7 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
                 CatchBlock catchStmt = new CatchBlock(vo,catchExecStmt); 
                 tryCatchBlocks.add(catchStmt);
                 this.methodCtx.returned = this.methodCtx.returned && tryReturned;
-                this.popFrame();
+                methodCtx.popFrame();
             }
         }
         BlockStmt tryFinallyStmt = null;
@@ -1644,14 +1453,14 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
     }
     
     public void visitBlockStmt(StatContext[] stats,BlockStmt blockStmt){
-        this.newFrame();
+        methodCtx.newFrame();
         if (stats == null) {
             return;
         }
         for (StatContext s : stats) {
             blockStmt.statements.add(visitStat(s));
         }
-        this.popFrame();
+        methodCtx.popFrame();
     }
 
     @Override
@@ -1910,12 +1719,12 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
         if(var==null) throw Exceptions.unexceptedValue(var);
         return var;
     }
-    
+
     @Nullable
     private LocalVarNode declareLocalVar(String name,Type type,int modifier,ParserRuleContext ctx){
         LocalVarNode localVarNode = new LocalVarNode(type,name,modifier);
-        ParameterNode param = this.getNamedParameter(name);
-        LocalVarNode var = this.getNamedLocalVar(name);
+        ParameterNode param = methodCtx.getNamedParameter(name);
+        LocalVarNode var = methodCtx.getNamedLocalVar(name);
         if(param!=null || var!=null){
             handleSyntaxError("variable is defined", ctx);
             return null;
@@ -2075,30 +1884,6 @@ public class AstBuilder extends AstBuilderBase implements KalangParserVisitor<Ob
         ExprNode arrExpr = createInitializedArray(type, initExprs);
         mapAst(arrExpr, ctx);
         return arrExpr;
-    }
-
-    private Type getVarObjectType(VarObject p) {
-        Type type = this.methodCtx.overrideTypes.get(p);
-        if(type==null) type = p.getType();
-        //TODO handle other object type
-        if(type instanceof ClassType){
-            Integer ns = methodCtx.nullState.get(p);
-            NullableKind nullable;
-            if(ns==null){
-                nullable = ((ObjectType) type).getNullable();
-            }else if(ns==NULLSTATE_MUST_NONNULL){
-                nullable = NullableKind.NONNULL;
-            }else if(ns==NULLSTATE_UNKNOWN){
-                nullable = NullableKind.UNKNOWN;
-            }else if(ns==NULLSTATE_MUST_NULL|| ns== NULLSTATE_NULLABLE){
-                nullable = NullableKind.NULLABLE;
-            }else{
-                throw Exceptions.unexceptedValue(ns);
-            }
-            return Types.getClassType((ClassType)type,nullable);
-        }else{
-            return type;
-        }
     }
 
     @Override
